添加发布到github脚本

+++ deploy-github.sh
@@ -1,12 +1,18 @@
 #!/bin/sh
 
+if [[ $(git status -s) ]]
+then
+    echo "The working directory is dirty. Please commit any pending changes."
+    exit 1;
+fi
+
 echo "Deleting old publication"
 rm -rf public
 mkdir public
 rm -rf .git/worktrees/public/
 
 echo "Checking out gh-pages branch into public"
-git worktree add -B gh-pages public origin/gh-pages
+git worktree add -B gh-pages public github/gh-pages
 
 echo "Removing existing files"
 rm -rf public/*
@@ -17,5 +23,5 @@ hugo
 echo "Updating gh-pages branch"
 cd public && git add --all && git commit -m "Publishing to gh-pages (publish.sh)"
 
-echo "Push to origin"
-git push origin gh-pages
+echo "Push to github"
+git push github gh-pages
